use VSSetup to discover msbuild

+++ src/Build.ps1
@@ -4,6 +4,7 @@ properties {
     $nugetBin="$PSScriptRoot\..\bin\nuget\"
     $version="2.0.0.0"
     $nuspecFiles=Get-ChildItem $PSScriptRoot *.nuspec -Recurse
+    $msbuild=FindMSBuild
 }
 
 task ChangeAssemblyInfo {
@@ -12,14 +13,15 @@ task ChangeAssemblyInfo {
 
 task Compile {
     exec {
-        # & msbuild .\PocketXaf.sln /t:Clean /fl
-        # if (! $?) { throw "msbuild failed" }
+        & $msbuild .\PocketXaf.sln /fl
+        if (! $?) { throw "compile failed" }
     }
 }
 
 task Clean{
     exec{
-        # & msbuild .\PocketXaf.sln /fl
+        & $msbuild .\PocketXaf.sln /t:Clean /fl
+        if (! $?) { throw "clean failed" }
         if (Test-Path $nugetBin){
             Remove-Item $nugetBin -Force -Recurse
         }
@@ -71,13 +73,16 @@ Task PackNuspec{
     }
 }
 
-task default  -depends ShowMsBuildVersion ,Clean ,Compile ,UpdateNuspecMetadata,PackNuspec
+task default  -depends Clean ,Compile ,UpdateNuspecMetadata,PackNuspec
 
-task ShowMsBuildVersion {
-    exec{
-        msbuild /version
-        Set-Location $PSScriptRoot
+function FindMSBuild() {
+    if (!(Get-Module -ListAvailable -Name VSSetup)) {
+        Write-Host "Module exists"
+        Set-PSRepository -Name "PSGallery" -InstallationPolicy Trusted
+        Install-Module VSSetup
     }
+    $path=Get-VSSetupInstance  | Select-VSSetupInstance -Product Microsoft.VisualStudio.Product.BuildTools -Latest |Select-Object -ExpandProperty InstallationPath
+    return join-path $path MSBuild\15.0\Bin\MSBuild.exe
 }
 
 task ? -Description "Helper to display task info" {
